feat: remove logs, create form component
