+++ src/components/project/pokebro-map/addNewMarkDialog.tsx
@@ -1,6 +1,4 @@
-import { useRouter } from "next/navigation";
-import { FieldValues, useForm } from "react-hook-form";
-import { type ReactNode } from "react";
+import { type ReactNode, useState } from "react";
 import { z } from "zod";
 
 import {
@@ -16,21 +14,7 @@ import {
     ContextMenuItem,
     ContextMenuTrigger,
 } from "@/components/ui/context-menu";
-import FormItemRender, { FormItemRenderProps } from "@/components/form/formItemRender";
-import { zodResolver } from "@hookform/resolvers/zod";
-import { Button } from "@/components/ui/button";
-import { Input } from "@/components/ui/input";
-import { Form } from "@/components/ui/form";
-import { api } from "@/trpc/react";
-
-const formSchema = z.object({
-    name: z.string().min(3),
-    posX: z.coerce.number(),
-    posY: z.coerce.number(),
-    floor: z.coerce.number(),
-    information: z.string(),
-    type: z.string(),
-});
+import AddNewMarkDialogForm from "@/components/project/pokebro-map/addNewMarkDialogForm";
 
 type addNewMarkDialogProps = {
     children: ReactNode;
@@ -38,31 +22,11 @@ type addNewMarkDialogProps = {
 };
 
 export default function AddNewMarkDialog({ children, getMousePosition }: addNewMarkDialogProps) {
-    const form = useForm<z.infer<typeof formSchema>>({
-        resolver: zodResolver(formSchema),
-        defaultValues: {
-            name: "",
-            posX: 0,
-            posY: 0,
-            floor: 0,
-            information: "",
-            type: "",
-        },
-    });
-
-    const markCreate = api.pokebroMap.createMarker.useMutation();
-    const router = useRouter();
+    const [clickPosition, setClickPosition] = useState({ x: 0, y: 0, floor: 0 });
 
     function handleMarkClick(e: MouseEvent) {
         const { x, y, floor } = getMousePosition(e);
-        form.setValue("posX", x);
-        form.setValue("posY", y);
-        form.setValue("floor", floor);
-    }
-
-    function onSubmit(values: z.infer<typeof formSchema>) {
-        markCreate.mutate(values);
-        router.refresh();
+        setClickPosition({ x, y, floor });
     }
 
     return (
@@ -85,52 +49,7 @@ export default function AddNewMarkDialog({ children, getMousePosition }: addNewM
                     <DialogHeader>
                         <DialogTitle>New marker</DialogTitle>
                     </DialogHeader>
-                    <Form {...form}>
-                        <form onSubmit={form.handleSubmit(onSubmit)}>
-                            <FormItemRender
-                                form={form}
-                                name={"name"}
-                                label={"Name"}
-                                renderItem={(field) => <Input {...field} />}
-                            />
-                            <FormItemRender
-                                form={form}
-                                name={"posX"}
-                                label={"Position X"}
-                                renderItem={(field) => <Input {...field} type={"number"} />}
-                            />
-                            <FormItemRender
-                                form={form}
-                                name={"posY"}
-                                label={"Position Y"}
-                                renderItem={(field) => <Input {...field} type={"number"} />}
-                            />
-                            <FormItemRender
-                                form={form}
-                                name={"floor"}
-                                label={"Floor"}
-                                renderItem={(field) => <Input {...field} type={"number"} />}
-                            />
-                            <FormItemRender
-                                form={form}
-                                name={"information"}
-                                label={"Information"}
-                                renderItem={(field) => <Input {...field} />}
-                            />
-                            <FormItemRender
-                                form={form}
-                                name={"type"}
-                                label={"Type"}
-                                renderItem={(field) => <Input {...field} />}
-                            />
-
-                            <div className={"flex justify-end pt-4"}>
-                                <DialogTrigger asChild>
-                                    <Button type="submit">Save changes</Button>
-                                </DialogTrigger>
-                            </div>
-                        </form>
-                    </Form>
+                    <AddNewMarkDialogForm clickPosition={clickPosition} />
                 </DialogContent>
             </Dialog>
         </>
